feat(admin): fix QR PDF generation, scanner status, and UI polish
