+++ frontend/src/app/models/checkInResponse.ts
@@ -1,5 +1,9 @@
 export interface CheckInResponse {
-  message: string;
+  message?: string;
   status: string;
   ticketId: number;
+  eventId?: number;
+  eventName?: string;
+  seatLabel?: string;
+  checkedInAt?: string;
 }
