+++ frontend/src/app/pages/admin-event-details/admin-event-details.html
@@ -1,0 +1,113 @@
+<section class="shell event-details-page animate-fade-in" *ngIf="event">
+  
+  <a routerLink="/admin" class="back-link">
+    <svg viewBox="0 0 24 24" width="20" height="20" stroke="currentColor" stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round" style="margin-right: 6px;">
+      <line x1="19" y1="12" x2="5" y2="12"></line>
+      <polyline points="12 19 5 12 12 5"></polyline>
+    </svg>
+    Назад к событиям
+  </a>
+
+  <div class="event-header">
+    <h1>{{ event.title }}</h1>
+    <div class="event-meta">
+      <span class="date">{{ event.startTime | date:'d MMMM HH:mm' }}</span>
+      <span class="venue" *ngIf="event.venue?.name">• {{ event.venue?.name }}</span>
+    </div>
+  </div>
+
+  <div class="action-grid animate-slide-up">
+    <button class="action-card qr-card" (click)="startScanner()">
+      <div class="action-icon">
+        <svg viewBox="0 0 24 24" width="48" height="48" stroke="currentColor" stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+          <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
+          <rect x="7" y="7" width="3" height="3"></rect>
+          <rect x="14" y="7" width="3" height="3"></rect>
+          <rect x="7" y="14" width="3" height="3"></rect>
+          <rect x="14" y="14" width="3" height="3"></rect>
+        </svg>
+      </div>
+      <h3>QR Сканер</h3>
+      <p>Быстрая проверка по QR-коду</p>
+    </button>
+
+    <button class="action-card manual-card" (click)="openCheckIn()">
+      <div class="action-icon">
+        <svg viewBox="0 0 24 24" width="48" height="48" stroke="currentColor" stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
+          <path d="M4 19.5A2.5 2.5 0 0 1 6.5 17H20"></path>
+          <path d="M6.5 2H20v20H6.5A2.5 2.5 0 0 1 4 19.5v-15A2.5 2.5 0 0 1 6.5 2z"></path>
+        </svg>
+      </div>
+      <h3>Проверка ID</h3>
+      <p>Ввод номера билета вручную</p>
+    </button>
+  </div>
+
+  <!-- Modal for Manual ID Check -->
+  <div class="created-modal" *ngIf="showCheckIn">
+    <div class="created-card scanner-card">
+      <h2>Ручная проверка билета</h2>
+      
+      <div class="input-group">
+        <label>ID Билета</label>
+        <input type="number" [(ngModel)]="ticketId" placeholder="Например: 12345" />
+      </div>
+      
+      <button class="primary-btn" (click)="checkInById()" [disabled]="!ticketId || loading">
+        {{ loading ? 'Проверка...' : 'Проверить' }}
+      </button>
+
+      <div class="scanner-container" *ngIf="checkInResult" style="margin-top: 16px; border: none; min-height: 180px; height: auto;">
+        <div class="scanner-overlay" [ngClass]="checkInResult.status === 'USED' ? 'success' : 'error'" style="position: relative; border-radius: 12px; animation: none;">
+          <div class="overlay-icon">
+            <svg *ngIf="checkInResult.status === 'USED'" viewBox="0 0 24 24" width="48" height="48" stroke="currentColor" stroke-width="2" fill="none">
+              <polyline points="20 6 9 17 4 12"></polyline>
+            </svg>
+            <svg *ngIf="checkInResult.status !== 'USED'" viewBox="0 0 24 24" width="48" height="48" stroke="currentColor" stroke-width="2" fill="none">
+              <circle cx="12" cy="12" r="10"></circle>
+              <line x1="15" y1="9" x2="9" y2="15"></line>
+              <line x1="9" y1="9" x2="15" y2="15"></line>
+            </svg>
+          </div>
+          <h3 class="overlay-title" style="font-size: 1.4rem;">
+            {{ checkInResult.status === 'USED' ? 'УСПЕШНО' : (checkInResult.status === 'ALREADY_USED' ? 'УЖЕ ПОГАШЕН' : 'ОШИБКА') }}
+          </h3>
+          <p class="overlay-message" style="font-size: 0.9rem;">{{ checkInMessage }}</p>
+        </div>
+      </div>
+
+      <button class="secondary-btn" (click)="closeCheckIn()" style="margin-top: auto;">Закрыть</button>
+    </div>
+  </div>
+
+  <!-- Modal for QR Scanner -->
+  <div class="created-modal" *ngIf="showScanner">
+    <div class="created-card scanner-card">
+      <h2>QR Scanner</h2>
+
+      <div class="scanner-container">
+        <div id="qr-reader" style="width: 100%"></div>
+        
+        <div class="scanner-overlay" *ngIf="checkInResult" [ngClass]="checkInResult.status === 'USED' ? 'success' : 'error'">
+          <div class="overlay-icon">
+            <svg *ngIf="checkInResult.status === 'USED'" viewBox="0 0 24 24" width="64" height="64" stroke="currentColor" stroke-width="2" fill="none">
+              <polyline points="20 6 9 17 4 12"></polyline>
+            </svg>
+            <svg *ngIf="checkInResult.status !== 'USED'" viewBox="0 0 24 24" width="64" height="64" stroke="currentColor" stroke-width="2" fill="none">
+              <circle cx="12" cy="12" r="10"></circle>
+              <line x1="15" y1="9" x2="9" y2="15"></line>
+              <line x1="9" y1="9" x2="15" y2="15"></line>
+            </svg>
+          </div>
+          <h3 class="overlay-title">
+            {{ checkInResult.status === 'USED' ? 'УСПЕШНО' : (checkInResult.status === 'ALREADY_USED' ? 'УЖЕ ПОГАШЕН' : 'ОШИБКА') }}
+          </h3>
+          <p class="overlay-message">{{ checkInMessage }}</p>
+        </div>
+      </div>
+
+      <button class="secondary-btn" (click)="closeScanner()">Закрыть</button>
+    </div>
+  </div>
+
+</section>
