+++ frontend/src/app/pages/admin/admin.component.css
@@ -65,6 +65,7 @@
   border: 1px solid rgba(255, 255, 255, 0.05);
   backdrop-filter: blur(10px);
   transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  cursor: pointer;
 }
 
 .admin-row:hover {
