+++ frontend/src/app/pages/admin/admin.component.html
@@ -1,7 +1,4 @@
 <section class="shell admin-page">
-  <div class="notification" *ngIf="notification">
-    {{ notification }}
-  </div>
   <div class="admin-top animate-fade-in">
     <h1>Ваши события</h1>
     <a routerLink="/admin/create" class="action-chip" style="font-size: 1rem; padding: 0.75rem 1.25rem;">
@@ -11,22 +8,15 @@
       </svg>
       Добавить событие
     </a>
-    <button class="action-chip" (click)="openCheckIn()">
-      Проверка id билета
-    </button>
-
-    <button class="action-chip" (click)="startScanner()">
-      QR Сканер
-    </button>
   </div>
   <div class="admin-list animate-slide-up">
-    <div class="admin-row" *ngFor="let event of events">
+    <div class="admin-row clickable-card" *ngFor="let event of events" [routerLink]="['/admin/events', event.id]">
       <div class="event-main">
         <h2>{{ event.title }}</h2>
         <span>{{ event.nextSession| date:'d MMMM HH:mm' }}</span>
       </div>
 
-      <div class="admin-actions">
+      <div class="admin-actions" (click)="$event.stopPropagation()">
         <button class="icon-btn edit" [routerLink]="['/admin/create', event.id]" title="Редактировать">
           <svg viewBox="0 0 24 24" width="18" height="18" stroke="currentColor" stroke-width="2" fill="none" stroke-linecap="round" stroke-linejoin="round">
             <path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"></path>
@@ -41,68 +31,6 @@
           </svg>
         </button>
       </div>
-    </div>
-  </div>
-
-  <div class="created-modal" *ngIf="showCreatedModal">
-    <div class="created-card">
-      <h3>Событие добавлено</h3>
-      <button type="button">Закрыть</button>
-    </div>
-  </div>
-  <div class="created-modal" *ngIf="showCheckIn">
-    <div class="created-card">
-
-      <h2>Check-in</h2>
-
-      <h4>По ID билета</h4>
-      <input type="number" [(ngModel)]="ticketId" placeholder="Ticket ID" />
-      <button (click)="checkInById()">Подтвердить</button>
-
-      <hr />
-
-      <h4>По QR токену</h4>
-      <input type="text" [(ngModel)]="token" placeholder="QR token" />
-      <button (click)="checkInByToken()">Сканировать</button>
-
-      <hr />
-
-      <div *ngIf="checkInMessage" class="checkin-result">
-        <pre>{{ checkInMessage }}</pre>
-      </div>
-
-      <button (click)="closeCheckIn()">Закрыть</button>
-
-    </div>
-  </div>
-  <div class="created-modal" *ngIf="showScanner">
-    <div class="created-card scanner-card">
-
-      <h2>QR Scanner</h2>
-
-      <div class="scanner-container">
-        <div id="qr-reader" style="width: 100%"></div>
-        
-        <div class="scanner-overlay" *ngIf="checkInResult" [ngClass]="checkInResult.status === 'SUCCESS' ? 'success' : 'error'">
-          <div class="overlay-icon">
-            <svg *ngIf="checkInResult.status === 'SUCCESS'" viewBox="0 0 24 24" width="64" height="64" stroke="currentColor" stroke-width="2" fill="none">
-              <polyline points="20 6 9 17 4 12"></polyline>
-            </svg>
-            <svg *ngIf="checkInResult.status !== 'SUCCESS'" viewBox="0 0 24 24" width="64" height="64" stroke="currentColor" stroke-width="2" fill="none">
-              <circle cx="12" cy="12" r="10"></circle>
-              <line x1="15" y1="9" x2="9" y2="15"></line>
-              <line x1="9" y1="9" x2="15" y2="15"></line>
-            </svg>
-          </div>
-          <h3 class="overlay-title">
-            {{ checkInResult.status === 'SUCCESS' ? 'УСПЕШНО' : (checkInResult.status === 'ALREADY_USED' ? 'УЖЕ ПОГАШЕН' : 'ОШИБКА') }}
-          </h3>
-          <p class="overlay-message">{{ checkInMessage }}</p>
-        </div>
-      </div>
-
-      <button (click)="closeScanner()">Закрыть</button>
-
     </div>
   </div>
 </section>
