+++ frontend/src/app/pages/profile/profile.component.css
@@ -277,11 +277,10 @@
 }
 
 .qr-label {
-  font-family: var(--font-family-body);
-  font-size: 0.75rem;
+  font-size: 0.88rem;
   font-weight: 700;
-  text-transform: uppercase;
-  letter-spacing: 0.3px;
+  color: #1a1b1e;
+  margin-top: 4px;
 }
 
 .download-pdf-btn {
