+++ frontend/src/app/pages/profile/profile.component.html
@@ -69,9 +69,6 @@
           <button class="download-pdf-btn" (click)="downloadTicketPDF(booking.ticketId)">
             Скачать PDF
           </button>
-          <button class="download-pdf-btn" (click)="previewPdf(booking.ticketId)">
-            Просмотр билета
-          </button>
         </div>
 
       </div>
